Reconcile and Merge Members

build_messages() reconciles and merges the members found in im_orig_dfs$orig_members and im_core_dfs$core_members.

build_members(
  summarize = TRUE,
  split = FALSE,
  as_tibble = ironmarch_as_tibble()
)

Arguments

summarize
logical(1L), Default: TRUE. Whether to aggregate all observations into single rows, prioritizing values found in im_core_dfs.

split
logical(1L), Default: FALSE. Whether to split the result into a named list of separate data frames, retaining all observations. Ignored if summarize is TRUE.

as_tibble
logical(1L), Default: TRUE. Whether a tibble::tibble() should be returned. Ignored if the {tibble} package is not installed.

Value

Data Frame

default: a tibble::tibble() if as_tibble is TRUE and {tibble} is installed.

alternative: a data.table::data.table()

If split is TRUE and summarize is FALSE, the result will be a list() of data frames.

Details

Values found in im_core_dfs$core_members are preferred. Otherwise values found in im_orig_dfs$orig_members are kept.

Examples

# members summarized to single rows ===================================================
build_members()

#> # A tibble: 1,542 x 83
#>    member_id name  member_group_id email joined              ip_address
#>        <int> <chr>           <int> <chr> <dttm>              <chr>     
#>  1         1 Алек…               4 slav… 2011-09-12 15:46:59 178.140.1…
#>  2         2 Phal…              13 illu… 2011-09-12 15:52:30 68.37.21.…
#>  3         3 Bloo…              13 rene… 2011-09-12 18:10:11 68.10.255…
#>  4         4 Mier…              13 homi… 2011-09-12 20:27:35 82.29.169…
#>  5         5 Will…              14 tash… 2011-09-12 20:58:17 90.214.15…
#>  6         7 Dadd…              13 benj… 2011-09-13 01:22:58 81.141.31…
#>  7         8 Mr. …              13 elei… 2011-09-13 04:24:07 98.198.24…
#>  8         9 Woma…              14 sol.… 2011-09-13 10:27:52 88.147.27…
#>  9        11 Four…              13 o244… 2011-09-13 12:14:28 109.77.91…
#> 10        12 Jami…              14 jami… 2011-09-13 17:43:02 86.134.81…
#> # … with 1,532 more rows, and 77 more variables: allow_admin_mails <lgl>,
#> #   skin <int>, warn_level <int>, warn_lastwarn <int>, restrict_post <lgl>,
#> #   bday_day <int>, bday_month <int>, bday_year <int>, msg_count_new <int>,
#> #   msg_count_total <int>, msg_show_notification <lgl>, last_visit <dttm>,
#> #   last_activity <dttm>, mod_posts <lgl>, auto_track <chr>, temp_ban <lgl>,
#> #   mgroup_others <chr>, member_login_key_expire <dttm>,
#> #   members_seo_name <chr>, members_cache <chr>, members_disable_pm <int>,
#> #   failed_logins <chr>, members_profile_views <int>, members_pass_hash <chr>,
#> #   members_pass_salt <chr>, members_bitoptions <int>, members_day_posts <chr>,
#> #   notification_cnt <int>, posts <int>, title <chr>, time_offset <chr>,
#> #   last_post <dttm>, view_sigs <lgl>, msg_count_reset <lgl>, dst_in_use <lgl>,
#> #   login_anonymous <chr>, ignored_users <chr>, org_perm_id <chr>,
#> #   member_login_key <chr>, members_auto_dst <lgl>, members_display_name <chr>,
#> #   members_l_display_name <chr>, members_l_username <chr>,
#> #   member_banned <lgl>, member_uploader <chr>, language <int>,
#> #   failed_login_count <int>, pp_last_visitors <chr>, pp_main_photo <chr>,
#> #   pp_main_width <int>, pp_main_height <int>, pp_thumb_photo <chr>,
#> #   pp_thumb_width <int>, pp_thumb_height <int>,
#> #   pp_setting_count_comments <lgl>, pp_reputation_points <int>,
#> #   pp_gravatar <chr>, pp_photo_type <chr>, signature <chr>,
#> #   fb_bwoptions <lgl>, tc_last_sid_import <chr>, tc_bwoptions <lgl>,
#> #   pp_customization <chr>, timezone <chr>, pp_cover_photo <chr>,
#> #   members_bitoptions2 <int>, create_menu <chr>, marked_site_read <int>,
#> #   pp_cover_offset <int>, acp_skin <lgl>, acp_language <lgl>,
#> #   member_title <chr>, member_posts <int>, member_last_post <int>,
#> #   member_streams <chr>, photo_last_update <dttm>, msg_count_reset_at <dttm>

# all member observations =============================================================
build_members(summarize = FALSE)

#> # A tibble: 1,970 x 84
#>    which_df member_id name  member_group_id email joined              ip_address
#>    <chr>        <int> <chr>           <int> <chr> <dttm>              <chr>     
#>  1 core             1 Алек…               4 slav… 2011-09-12 15:46:59 178.140.1…
#>  2 core             2 Phal…              13 illu… 2011-09-12 15:52:30 68.37.21.…
#>  3 core             3 Bloo…              13 rene… 2011-09-12 18:10:11 68.10.255…
#>  4 core             4 Mier…              13 homi… 2011-09-12 20:27:35 82.29.169…
#>  5 core             5 Will…              14 tash… 2011-09-12 20:58:17 90.214.15…
#>  6 core             7 Dadd…              13 benj… 2011-09-13 01:22:58 81.141.31…
#>  7 core             8 Mr. …              13 elei… 2011-09-13 04:24:07 98.198.24…
#>  8 core             9 Woma…              14 sol.… 2011-09-13 10:27:52 88.147.27…
#>  9 core            11 Four…              13 o244… 2011-09-13 12:14:28 109.77.91…
#> 10 core            12 Jami…              14 jami… 2011-09-13 17:43:02 86.134.81…
#> # … with 1,960 more rows, and 77 more variables: allow_admin_mails <lgl>,
#> #   skin <int>, warn_level <int>, warn_lastwarn <int>, restrict_post <lgl>,
#> #   bday_day <int>, bday_month <int>, bday_year <int>, msg_count_new <int>,
#> #   msg_count_total <int>, msg_show_notification <lgl>, last_visit <dttm>,
#> #   last_activity <dttm>, mod_posts <lgl>, auto_track <chr>, temp_ban <lgl>,
#> #   mgroup_others <chr>, member_login_key_expire <dttm>,
#> #   members_seo_name <chr>, members_cache <chr>, members_disable_pm <int>,
#> #   failed_logins <chr>, members_profile_views <int>, members_pass_hash <chr>,
#> #   members_pass_salt <chr>, members_bitoptions <int>, members_day_posts <chr>,
#> #   notification_cnt <int>, posts <int>, title <chr>, time_offset <chr>,
#> #   last_post <dttm>, view_sigs <lgl>, msg_count_reset <lgl>, dst_in_use <lgl>,
#> #   login_anonymous <chr>, ignored_users <chr>, org_perm_id <chr>,
#> #   member_login_key <chr>, members_auto_dst <lgl>, members_display_name <chr>,
#> #   members_l_display_name <chr>, members_l_username <chr>,
#> #   member_banned <lgl>, member_uploader <chr>, language <int>,
#> #   failed_login_count <int>, pp_last_visitors <chr>, pp_main_photo <chr>,
#> #   pp_main_width <int>, pp_main_height <int>, pp_thumb_photo <chr>,
#> #   pp_thumb_width <int>, pp_thumb_height <int>,
#> #   pp_setting_count_comments <lgl>, pp_reputation_points <int>,
#> #   pp_gravatar <chr>, pp_photo_type <chr>, signature <chr>,
#> #   fb_bwoptions <lgl>, tc_last_sid_import <chr>, tc_bwoptions <lgl>,
#> #   pp_customization <chr>, timezone <chr>, pp_cover_photo <chr>,
#> #   members_bitoptions2 <int>, create_menu <chr>, marked_site_read <int>,
#> #   pp_cover_offset <int>, acp_skin <lgl>, acp_language <lgl>,
#> #   member_title <chr>, member_posts <int>, member_last_post <int>,
#> #   member_streams <chr>, photo_last_update <dttm>, msg_count_reset_at <dttm>

# all member observations split into a list of data frames ============================
split_members <- build_members(summarize = FALSE, split = TRUE)

# first three members that appear in both... ==========================================
# ... `im_orig_dfs$orig_members` and `im_core_dfs$core_members`
split_members[
  vapply(split_members, function(.x) nrow(.x) == 2, logical(1L))
][1:3]

#> $`1`
#> # A tibble: 2 x 84
#>   which_df member_id name  member_group_id email joined              ip_address
#>   <chr>        <int> <chr>           <int> <chr> <dttm>              <chr>     
#> 1 core             1 Алек…               4 slav… 2011-09-12 15:46:59 178.140.1…
#> 2 orig             1 Alex…               4 slav… 2011-09-12 15:46:59 178.140.1…
#> # … with 77 more variables: allow_admin_mails <lgl>, skin <int>,
#> #   warn_level <int>, warn_lastwarn <int>, restrict_post <lgl>, bday_day <int>,
#> #   bday_month <int>, bday_year <int>, msg_count_new <int>,
#> #   msg_count_total <int>, msg_show_notification <lgl>, last_visit <dttm>,
#> #   last_activity <dttm>, mod_posts <lgl>, auto_track <chr>, temp_ban <lgl>,
#> #   mgroup_others <chr>, member_login_key_expire <dttm>,
#> #   members_seo_name <chr>, members_cache <chr>, members_disable_pm <int>,
#> #   failed_logins <chr>, members_profile_views <int>, members_pass_hash <chr>,
#> #   members_pass_salt <chr>, members_bitoptions <int>, members_day_posts <chr>,
#> #   notification_cnt <int>, posts <int>, title <chr>, time_offset <chr>,
#> #   last_post <dttm>, view_sigs <lgl>, msg_count_reset <lgl>, dst_in_use <lgl>,
#> #   login_anonymous <chr>, ignored_users <chr>, org_perm_id <chr>,
#> #   member_login_key <chr>, members_auto_dst <lgl>, members_display_name <chr>,
#> #   members_l_display_name <chr>, members_l_username <chr>,
#> #   member_banned <lgl>, member_uploader <chr>, language <int>,
#> #   failed_login_count <int>, pp_last_visitors <chr>, pp_main_photo <chr>,
#> #   pp_main_width <int>, pp_main_height <int>, pp_thumb_photo <chr>,
#> #   pp_thumb_width <int>, pp_thumb_height <int>,
#> #   pp_setting_count_comments <lgl>, pp_reputation_points <int>,
#> #   pp_gravatar <chr>, pp_photo_type <chr>, signature <chr>,
#> #   fb_bwoptions <lgl>, tc_last_sid_import <chr>, tc_bwoptions <lgl>,
#> #   pp_customization <chr>, timezone <chr>, pp_cover_photo <chr>,
#> #   members_bitoptions2 <int>, create_menu <chr>, marked_site_read <int>,
#> #   pp_cover_offset <int>, acp_skin <lgl>, acp_language <lgl>,
#> #   member_title <chr>, member_posts <int>, member_last_post <int>,
#> #   member_streams <chr>, photo_last_update <dttm>, msg_count_reset_at <dttm>
#> 
#> $`2`
#> # A tibble: 2 x 84
#>   which_df member_id name  member_group_id email joined              ip_address
#>   <chr>        <int> <chr>           <int> <chr> <dttm>              <chr>     
#> 1 core             2 Phal…              13 illu… 2011-09-12 15:52:30 68.37.21.…
#> 2 orig             2 Kacen              13 illu… 2011-09-12 15:52:30 68.37.21.…
#> # … with 77 more variables: allow_admin_mails <lgl>, skin <int>,
#> #   warn_level <int>, warn_lastwarn <int>, restrict_post <lgl>, bday_day <int>,
#> #   bday_month <int>, bday_year <int>, msg_count_new <int>,
#> #   msg_count_total <int>, msg_show_notification <lgl>, last_visit <dttm>,
#> #   last_activity <dttm>, mod_posts <lgl>, auto_track <chr>, temp_ban <lgl>,
#> #   mgroup_others <chr>, member_login_key_expire <dttm>,
#> #   members_seo_name <chr>, members_cache <chr>, members_disable_pm <int>,
#> #   failed_logins <chr>, members_profile_views <int>, members_pass_hash <chr>,
#> #   members_pass_salt <chr>, members_bitoptions <int>, members_day_posts <chr>,
#> #   notification_cnt <int>, posts <int>, title <chr>, time_offset <chr>,
#> #   last_post <dttm>, view_sigs <lgl>, msg_count_reset <lgl>, dst_in_use <lgl>,
#> #   login_anonymous <chr>, ignored_users <chr>, org_perm_id <chr>,
#> #   member_login_key <chr>, members_auto_dst <lgl>, members_display_name <chr>,
#> #   members_l_display_name <chr>, members_l_username <chr>,
#> #   member_banned <lgl>, member_uploader <chr>, language <int>,
#> #   failed_login_count <int>, pp_last_visitors <chr>, pp_main_photo <chr>,
#> #   pp_main_width <int>, pp_main_height <int>, pp_thumb_photo <chr>,
#> #   pp_thumb_width <int>, pp_thumb_height <int>,
#> #   pp_setting_count_comments <lgl>, pp_reputation_points <int>,
#> #   pp_gravatar <chr>, pp_photo_type <chr>, signature <chr>,
#> #   fb_bwoptions <lgl>, tc_last_sid_import <chr>, tc_bwoptions <lgl>,
#> #   pp_customization <chr>, timezone <chr>, pp_cover_photo <chr>,
#> #   members_bitoptions2 <int>, create_menu <chr>, marked_site_read <int>,
#> #   pp_cover_offset <int>, acp_skin <lgl>, acp_language <lgl>,
#> #   member_title <chr>, member_posts <int>, member_last_post <int>,
#> #   member_streams <chr>, photo_last_update <dttm>, msg_count_reset_at <dttm>
#> 
#> $`3`
#> # A tibble: 2 x 84
#>   which_df member_id name  member_group_id email joined              ip_address
#>   <chr>        <int> <chr>           <int> <chr> <dttm>              <chr>     
#> 1 core             3 Bloo…              13 rene… 2011-09-12 18:10:11 68.10.255…
#> 2 orig             3 Bloo…              13 rene… 2011-09-12 18:10:11 68.10.255…
#> # … with 77 more variables: allow_admin_mails <lgl>, skin <int>,
#> #   warn_level <int>, warn_lastwarn <int>, restrict_post <lgl>, bday_day <int>,
#> #   bday_month <int>, bday_year <int>, msg_count_new <int>,
#> #   msg_count_total <int>, msg_show_notification <lgl>, last_visit <dttm>,
#> #   last_activity <dttm>, mod_posts <lgl>, auto_track <chr>, temp_ban <lgl>,
#> #   mgroup_others <chr>, member_login_key_expire <dttm>,
#> #   members_seo_name <chr>, members_cache <chr>, members_disable_pm <int>,
#> #   failed_logins <chr>, members_profile_views <int>, members_pass_hash <chr>,
#> #   members_pass_salt <chr>, members_bitoptions <int>, members_day_posts <chr>,
#> #   notification_cnt <int>, posts <int>, title <chr>, time_offset <chr>,
#> #   last_post <dttm>, view_sigs <lgl>, msg_count_reset <lgl>, dst_in_use <lgl>,
#> #   login_anonymous <chr>, ignored_users <chr>, org_perm_id <chr>,
#> #   member_login_key <chr>, members_auto_dst <lgl>, members_display_name <chr>,
#> #   members_l_display_name <chr>, members_l_username <chr>,
#> #   member_banned <lgl>, member_uploader <chr>, language <int>,
#> #   failed_login_count <int>, pp_last_visitors <chr>, pp_main_photo <chr>,
#> #   pp_main_width <int>, pp_main_height <int>, pp_thumb_photo <chr>,
#> #   pp_thumb_width <int>, pp_thumb_height <int>,
#> #   pp_setting_count_comments <lgl>, pp_reputation_points <int>,
#> #   pp_gravatar <chr>, pp_photo_type <chr>, signature <chr>,
#> #   fb_bwoptions <lgl>, tc_last_sid_import <chr>, tc_bwoptions <lgl>,
#> #   pp_customization <chr>, timezone <chr>, pp_cover_photo <chr>,
#> #   members_bitoptions2 <int>, create_menu <chr>, marked_site_read <int>,
#> #   pp_cover_offset <int>, acp_skin <lgl>, acp_language <lgl>,
#> #   member_title <chr>, member_posts <int>, member_last_post <int>,
#> #   member_streams <chr>, photo_last_update <dttm>, msg_count_reset_at <dttm>
#>