Build Message Network

Build Message Network

im_message_network(edge_geo_distance = FALSE)

Arguments

edge_geo_distance
Whether to calculate dyadic geo-distances and add as an edge attribute named distance_km (for distance in kilometers). Member points are calculated from the centroids of each individual's geocoded IP addresses.

Examples

im_message_g <- im_message_network()

print(im_message_g, full = FALSE)

#> IGRAPH 070c571 U--- 821 41786 -- 
#> + attr: member_id (v/n), screen_name (v/c), member_group_id (v/n),
#> | email (v/c), ip_address (v/c), allow_admin_mails (v/l), skin (v/n),
#> | warn_level (v/n), warn_lastwarn (v/n), restrict_post (v/l), bday_day
#> | (v/n), bday_month (v/n), bday_year (v/n), msg_count_new (v/n),
#> | msg_count_total (v/n), msg_show_notification (v/l), mod_posts (v/l),
#> | auto_track (v/c), temp_ban (v/l), mgroup_others (v/c),
#> | members_seo_name (v/c), members_cache (v/c), members_disable_pm
#> | (v/n), failed_logins (v/c), members_profile_views (v/n),
#> | members_pass_hash (v/c), members_pass_salt (v/c), members_bitoptions
#> | (v/n), members_day_posts (v/c), notification_cnt (v/n), posts (v/n),
#> | title (v/c), time_offset (v/c), last_post (v/n), view_sigs (v/l),
#> | msg_count_reset (v/l), dst_in_use (v/l), login_anonymous (v/c),
#> | ignored_users (v/c), org_perm_id (v/c), member_login_key (v/c),
#> | members_auto_dst (v/l), members_display_name (v/c),
#> | members_l_display_name (v/c), members_l_username (v/c), member_banned
#> | (v/l), member_uploader (v/c), language (v/n), failed_login_count
#> | (v/n), pp_last_visitors (v/c), pp_main_photo (v/c), pp_main_width
#> | (v/n), pp_main_height (v/n), pp_thumb_photo (v/c), pp_thumb_width
#> | (v/n), pp_thumb_height (v/n), pp_setting_count_comments (v/l),
#> | pp_reputation_points (v/n), pp_gravatar (v/c), pp_photo_type (v/c),
#> | signature (v/c), fb_bwoptions (v/l), tc_last_sid_import (v/c),
#> | tc_bwoptions (v/l), pp_customization (v/c), timezone (v/c),
#> | pp_cover_photo (v/c), members_bitoptions2 (v/n), create_menu (v/c),
#> | marked_site_read (v/n), pp_cover_offset (v/n), acp_skin (v/l),
#> | acp_language (v/l), member_title (v/c), member_posts (v/n),
#> | member_last_post (v/n), member_streams (v/c), photo_last_update
#> | (v/n), msg_count_reset_at (v/n), source_msg_author_id (e/n),
#> | target_msg_author_id (e/n), msg_topic_id (e/n), msg_id (e/n),
#> | msg_date (e/n), msg_post (e/c), msg_author_id (e/n), msg_ip_address
#> | (e/c), msg_is_first_post (e/l)